Unify cubic field extension arithmetic
