+++ algebra/src/bls12_381/fields/fq6.rs
@@ -31,7 +31,7 @@ impl Fp6Parameters for Fq6Parameters {
     );
 
     #[rustfmt::skip]
-    const FROBENIUS_COEFF_FP6_C1: [Fq2; 6] = [
+    const FROBENIUS_COEFF_FP6_C1: &'static [Fq2] = &[
         // Fq2(u + 1)**(((q^0) - 1) / 3)
         field_new!(Fq2,
             field_new!(Fq, BigInteger384([
@@ -107,7 +107,7 @@ impl Fp6Parameters for Fq6Parameters {
     ];
 
     #[rustfmt::skip]
-    const FROBENIUS_COEFF_FP6_C2: [Fq2; 6] = [
+    const FROBENIUS_COEFF_FP6_C2: &'static [Fq2] = &[
         // Fq2(u + 1)**(((2q^0) - 2) / 3)
         field_new!(Fq2,
             field_new!(Fq, BigInteger384([
