+++ algebra/src/mnt6_298/fields/fq3.rs
@@ -55,7 +55,7 @@ impl Fp3Parameters for Fq3Parameters {
     );
 
     #[rustfmt::skip]
-    const FROBENIUS_COEFF_FP3_C1: [Fq; 3] = [
+    const FROBENIUS_COEFF_FP3_C1: &'static [Fq] = &[
         field_new!(Fq, BigInteger([
             0xc3177aefffbb845c,
             0x9b80c702f9961788,
@@ -80,7 +80,7 @@ impl Fp3Parameters for Fq3Parameters {
     ];
 
     #[rustfmt::skip]
-    const FROBENIUS_COEFF_FP3_C2: [Fq; 3] = [
+    const FROBENIUS_COEFF_FP3_C2: &'static [Fq] = &[
         field_new!(Fq, BigInteger([
             0xc3177aefffbb845c,
             0x9b80c702f9961788,
