+++ algebra/src/mnt6_753/fields/fq3.rs
@@ -95,7 +95,7 @@ impl Fp3Parameters for Fq3Parameters {
     // c1[1] = 24129022407817241407134263419936114379815707076943508280977368156625538709102831814843582780138963119807143081677569721953561801075623741378629346409604471234573396989178424163772589090105392407118197799904755622897541183052132
     // c1[2] = 17769468560101711995209951371304522748355002843010440790806134764399814103468274958215310983651375801610927890210888755369611256415970113691066895445191924931148019336171640277697829047741006062493737919155152541323243293107868,
     #[rustfmt::skip]
-    const FROBENIUS_COEFF_FP3_C1: [Fq; 3] = [
+    const FROBENIUS_COEFF_FP3_C1: &'static [Fq] = &[
         FQ_ONE,
         field_new!(Fq, BigInteger([
             7739145380395648640,
@@ -129,7 +129,7 @@ impl Fp3Parameters for Fq3Parameters {
 
     // c2 = {c1[0], c1[2], c1[1]}
     #[rustfmt::skip]
-    const FROBENIUS_COEFF_FP3_C2: [Fq; 3] = [
+    const FROBENIUS_COEFF_FP3_C2: &'static [Fq] = &[
         FQ_ONE,
         Self::FROBENIUS_COEFF_FP3_C1[2],
         Self::FROBENIUS_COEFF_FP3_C1[1],
